Refactor messages into ipc module
