--- astoria/common/messages/__init__.py
@@ -1,1 +1,0 @@
-"""MQTT Message Definitions."""
