+++ astoria/managers/astmetad/metadata_disk_lifecycle.py
@@ -11,7 +11,7 @@ import toml
 from pydantic.error_wrappers import ValidationError
 
 from astoria.common.config import SSID_PREFIX, AstoriaConfig, RobotSettings
-from astoria.common.messages.astdiskd import DiskInfo, DiskUUID
+from astoria.common.disks import DiskInfo, DiskUUID
 
 LOGGER = logging.getLogger(__name__)
