+++ tests/common/messages/test_astmetad.py
@@ -1,5 +1,6 @@
 """Tests for astmetad message definitions."""
-from astoria.common.messages.astmetad import Metadata, MetadataManagerMessage
+from astoria.common.ipc import MetadataManagerMessage
+from astoria.common.metadata import Metadata
 
 
 def test_metadata_manager_message_fields() -> None:
